Replace async-std with smol
