+++ atspi-proxies/Cargo.toml
@@ -26,7 +26,6 @@ serde        = { version = "^1.0.200", default-features = false, features = ["de
 zbus         = { workspace = true }
 
 [dev-dependencies]
-async-std    = { version = "1.13.0", features = ["attributes"] }
 atspi-common = { path = "../atspi-common", version = "0.11.0" }
 byteorder    = "1.4"
 futures-lite = { version = "2.6.0", default-features = false }
