+++ atspi/Cargo.toml
@@ -57,15 +57,14 @@ name = "focused-tokio"
 path = "./examples/focused-tokio.rs"
 
 [[example]]
-name = "focused-async-std"
-path = "./examples/focused-async-std.rs"
+name = "focused-smol"
+path = "./examples/focused-smol.rs"
 
 [[example]]
 name = "selected-text"
 path = "./examples/selected-text.rs"
 
 [dev-dependencies]
-async-std      = { version = "1.13", features = ["attributes"] }
 atspi          = { path = "." }
 # TOOD: pinned due to MSRV 1.77.2 policy; this can be lifted when we bump.
 criterion = "=0.5.0"
@@ -76,3 +75,4 @@ futures-lite   = { version = "2.6.0", default-features = false }
 tokio          = { version = "1.38.2", default-features = false, features = ["macros", "rt-multi-thread"] }
 tokio-stream   = "0.1.1"
 zbus.workspace = true
+smol-macros = "0.1.1"
